requirements.txt
matplotlib
numpy
pandas
scipy
scikit-learn
seaborn
tensorflow
keras
streamlit
pillow